requirements.txt
anthropic>=0.49.0
notion-client==2.2.1
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
